feat: move concatBytes to concat
