+++ packages/utils/src/utils/concat.ts
@@ -1,5 +1,5 @@
 import type { BytesLike } from 'ethers';
-import { getBytes } from 'ethers';
+import { getBytesCopy } from 'ethers';
 
 /**
  * Concatenates multiple Uint8Arrays into a single Uint8Array.
@@ -33,7 +33,7 @@ export const concatBytes = (
  * @returns - The concatenated array.
  */
 export const concat = (arrays: ReadonlyArray<BytesLike>): Uint8Array => {
-  const bytes = arrays.map((v) => getBytes(v));
+  const bytes = arrays.map((v) => getBytesCopy(v));
 
   return concatBytes(bytes);
 };
